Re-organise files to be more in line with recommendations.
I want to start writing a client api that people can use which will
share models used by the server it seems sensible to organise the code
in a more standard way first.
