+++ cmd/server/main.go
@@ -1,0 +1,7 @@
+package main
+
+import server "github.com/baelish/alive/internal/server"
+
+func main() {
+	server.Start()
+}
