+++ generate-new-static-files
@@ -12,5 +12,5 @@ if ! command -v go-bindata > /dev/null; then
     exit 1
 fi
 
-go-bindata -o static-files.go -prefix static-source static-source/...
+go-bindata -o internal/server/static-files.go -prefix static-source static-source/...
