+++ internal/server/create-static-content.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"os"
