+++ internal/server/df-functions.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"encoding/json"
