+++ internal/server/random-string.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"math/rand"
